refactor friend pages, to display different user actions
